Updated Default Services Updater Code
This is needed again to handle the new default services that are now without my hacky transform scale and have properly resized svg's. This allows the app to update the default services without the user interacting if it detects that they have not changed there configuration. If they changed thier config it gives then the same popup as the last release had.