Fixed Window Drag Functionality
The drag functionality never worked for me in development and so I only half implemented it. Someone reported problems due to it working but being in a half implemented state so this fixes that. Also changed the header and close btn cursor to be the grab cursor to match the service boxes on the menu.