Fixed Broken Netflix Player
Netflix recently pushed an update that doesn't support the old broswer version that ElectronPlayer was using. This update fixes the Netflix player. The fix for Mac and Linux are both different and incompatible so I have now moved to a dual package.json design. How to use can be seen in the Readme. The Linux version gets Electron 5.0.4 while the Mac version gets 4.2.4 but in the future I will work on bridging this gap and moving back to a single package.json file.