Changed Snap Confinement to Classic
Currently I need the app to work and this will do it. Something in the castlabs-electron (I think) is doing a sysctl call and the application is not able to start. Without castlabs-electron Netflix doesn't work. This is a temporary fix. I hate it but it will work until a better fix is developed.