fix(security): implemented final recommended security features
This includes adding a CSP to the menu interface and adding restriction so websites are not granted permissions to the users computer (camera, location, etc) without the user allowing it. The electron remote api's have also been locked down to prevent their malicious use by a website running inside the app. Changed the menu font to a websafe font so information isn't leaked to google when the font is fetched and also to make the CSP easier.