feat(config): A new config system
Instead of giving users an annoying popup and requiring a manual config update process on every update I have redone the config system to prevent this. It merges the user configs and the default config. This means if they never change the config (user config) the default config will always be used but can still be changed at any time by me.