fix: fixed error undefined mainWindow.webContents
This error happened when enabled and disabling menu options requiring a reload multiple times without a full restart. It was because of a cleanup function clearing the mainWindow when it shouldn't have. Reset all settings menu item now restarts the app not reload. This is so it doesn't remeber the window details as this is not what you want when resetting the config.