feat: option to open custom url from string + code cleanup
You can now open a custom url without it being permanently added as a service. This is good for a website with a video you just want to view once. The code has been cleaned up so instead of using mainWindow.webContents to interact with the url of the window mainWindow. is used.