fix: dep updates + Linux support for custom url prompt
The open custom URL prompt was made with a Mac only package so that
package has been replaced with a cross platform alternative. Lots of
deps were updated including electron on Linux and the Adblocker.
electron-store has been moved back to the master from my custom fork as
the issue's my fork fixed are fixed in the master.