feat: added analytics
This is done to make developing the app easier as I know which devices
to test on and it also allows me to know how many people are using the
app which is something I like to know. There is no option to disable
analytics but you could block esa.otbeaumont.me using your hosts file or
any other method if you are extremely concerned.